third_party/WebKit/LayoutTests/http/tests/xmlhttprequest/xmlhttprequest-big-document.html

<html>
<head>
<script type="text/javascript">

if (window.testRunner) {
    testRunner.dumpAsText();
    testRunner.waitUntilDone();
}

var xhr = new XMLHttpRequest();

xhr.onreadystatechange = function() {
    if (xhr.readyState == 4 && xhr.status == 200) {
        var doc = xhr.responseXML;
        var children = doc.getElementsByTagName("child");
        console.log(children);
        document.getElementById("page").textContent = "PASS";
        if (window.testRunner)
            testRunner.notifyDone();
    }
}

xhr.open("GET", "resources/big.xml", true);
xhr.send(null);
</script>
</head>
<body>
<p> Test case for <a href="https://code.google.com/p/chromium/issues/detail?id=528078">bug 528078</a>: XML files fail to parse if over 10 million bytes in size</p>
<p> You should see PASS.</p>
<div id="page"/>
</body>
</html>
