third_party/blink/web_tests/http/tests/xmlhttprequest/xmlhttprequest-big-document-expected.txt

CONSOLE MESSAGE: [object HTMLCollection]
Test case for bug 528078: XML files fail to parse if over 10 million bytes in size

You should see PASS.

PASS
